Rename ReservationFormContainer to ReservationCreateFormContainer (refs #171)
